fix(storage): Improve exception handling when attempting to overwrite a file during download

+++ aws-storage-s3/src/main/java/com/amplifyframework/storage/s3/transfer/TransferManager.kt
@@ -24,8 +24,11 @@ import aws.sdk.kotlin.services.s3.model.ObjectCannedAcl
 import com.amplifyframework.core.Amplify
 import com.amplifyframework.core.category.CategoryType
 import com.amplifyframework.storage.ObjectMetadata
+import com.amplifyframework.storage.StorageException
+import com.amplifyframework.storage.StorageFilePermissionException
 import com.amplifyframework.storage.s3.AWSS3StoragePlugin
 import com.amplifyframework.storage.s3.TransferOperations
+import com.amplifyframework.storage.s3.extensions.unableToOverwriteFileException
 import com.amplifyframework.storage.s3.transfer.worker.RouterWorker
 import com.amplifyframework.storage.s3.transfer.worker.TransferWorkerFactory
 import java.io.File
@@ -192,6 +195,9 @@ internal class TransferManager(
         val transferRecordId: Int = uri.lastPathSegment?.toInt()
             ?: throw IllegalStateException("Invalid TransferRecord ID ${uri.lastPathSegment}")
         if (file.isFile) {
+            if (!file.canWrite()) {
+                throw StorageFilePermissionException.unableToOverwriteFileException()
+            }
             logger.warn("Overwriting existing file: $file")
             file.delete()
         }
